# openapi.yaml
openapi: 3.0.3
info:
  title: Order Management System API
  description: REST API for managing customers, products, and orders
  version: 1.0.0
  contact:
    name: Interview Test

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: customers
    description: Customer management
  - name: products
    description: Product management
  - name: orders
    description: Order management

paths:
  /api/customers:
    get:
      tags:
        - customers
      summary: Get all customers
      description: Returns a list of all customers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags:
        - customers
      summary: Create a new customer
      description: Creates a new customer and returns the created entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /api/customers/{id}:
    get:
      tags:
        - customers
      summary: Get customer by ID
      description: Returns a single customer by their ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    put:
      tags:
        - customers
      summary: Update a customer
      description: Updates an existing customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    delete:
      tags:
        - customers
      summary: Delete a customer
      description: Deletes a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Customer deleted successfully

  /api/products:
    get:
      tags:
        - products
      summary: Get all products
      description: Returns a list of all products
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - products
      summary: Create a new product
      description: Creates a new product and returns the created entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /api/products/{id}:
    get:
      tags:
        - products
      summary: Get product by ID
      description: Returns a single product by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    put:
      tags:
        - products
      summary: Update a product
      description: Updates an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      tags:
        - products
      summary: Delete a product
      description: Deletes a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully

  /api/products/stock:
    put:
      tags:
        - products
      summary: Bulk update stock
      description: Updates stock for multiple products in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StockUpdateRequest'
      responses:
        '200':
          description: Stock updated successfully

  /api/orders:
    get:
      tags:
        - orders
      summary: Get paginated orders
      description: Returns a paginated list of orders
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
    post:
      tags:
        - orders
      summary: Create a new order
      description: Creates a new order and returns the created entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'

  /api/orders/all:
    get:
      tags:
        - orders
      summary: Get all orders
      description: Returns a list of all orders (non-paginated)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDTO'

  /api/orders/{id}:
    get:
      tags:
        - orders
      summary: Get order by ID
      description: Returns a single order by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '404':
          description: Order not found
    delete:
      tags:
        - orders
      summary: Delete an order (admin only)
      description: Deletes an order by ID. Requires ADMIN role.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Order deleted successfully
        '403':
          description: Forbidden - Admin role required

components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        createdAt:
          type: string
          format: date-time

    CustomerRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
        category:
          type: string

    ProductRequest:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
        category:
          type: string

    StockUpdateRequest:
      type: object
      required:
        - productId
        - quantityChange
      properties:
        productId:
          type: integer
          format: int64
        quantityChange:
          type: integer
          description: Positive to add stock, negative to remove

    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        orderDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - SHIPPED
            - DELIVERED
            - CANCELLED
        totalAmount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTO'

    OrderItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
        subtotal:
          type: number
          format: double

    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: integer
          format: int64
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1

    OrderPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean

    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        unpaged:
          type: boolean